Moving clone out of tests
